+++ changelog.md
@@ -10,6 +10,7 @@
 * Increase unit test coverage to >85%
 * Gracefully ignore missing author names/emails from conduit for summary badge
 * Report Cobertura coverage data to Harbormaster API
+* Add option to preserve formatting in additional comments
 
 ### 1.7.2 (2015/08/13)